requirements.txt
axi==0.1
cairocffi==1.1.0
cffi==1.14.0
click==7.1.1
decorator==4.4.2
Flask==1.1.2
Flask-API==2.0
gocept.cache==4.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.2
pkg-resources==0.0.0
pycparser==2.20
pyhull==2015.2.1
pyserial==3.4
Shapely==1.7.0
transaction==3.0.0
Werkzeug==1.0.1
zope.interface==5.1.0
zope.testing==4.7
